Split local workspace migration tests
